Refactor to reduce chains of dispatches to load.
