+++ client/src/init.js
@@ -1,55 +1,22 @@
 import store from './store'
 import watch from 'redux-watch'
-import {initialize, loadData} from './actions/FlowActions'
-import {decodeQueryString} from './utils/queryUtils'
 import {
-  triggerSearch,
-  fetchGranules,
-  clearCollections,
-  clearGranules,
-  getCollection,
-} from './actions/SearchRequestActions'
-import {
-  updateSearch,
-  clearSelections,
-  toggleSelection,
-} from './actions/SearchParamActions'
-import {
-  isDetailPage,
-  isGranuleListPage,
-  getCollectionIdFromDetailPath,
-  getCollectionIdFromGranuleListPath,
-} from './utils/urlUtils'
+  initialize,
+  loadDetails,
+  loadCollections,
+  loadGranulesList,
+} from './actions/FlowActions'
+import {isDetailPage, isGranuleListPage} from './utils/urlUtils'
 
 const loadFromUrl = (path, newQueryString) => {
-  // Note, collection queries are automatically updated by the URL because the query is parsed into search, which triggers loadData via a watch
-  if (
-    isDetailPage(path) &&
-    !store.getState().behavior.request.getCollectionInFlight
-  ) {
-    const detailId = getCollectionIdFromDetailPath(path)
-    store.dispatch(getCollection(detailId))
+  if (isDetailPage(path)) {
+    store.dispatch(loadDetails(path))
   }
   else if (isGranuleListPage(path)) {
-    const detailId = getCollectionIdFromGranuleListPath(path)
-    store.dispatch(clearSelections())
-    store.dispatch(toggleSelection(detailId))
-    store.dispatch(clearGranules())
-    store.dispatch(fetchGranules())
+    store.dispatch(loadGranulesList(path))
   }
   else {
-    if (newQueryString.indexOf('?') === 0) {
-      newQueryString = newQueryString.slice(1)
-    }
-    const searchFromQuery = decodeQueryString(newQueryString)
-    const searchFromState = _.get(store.getState(), 'behavior.search')
-    if (!_.isEqual(searchFromQuery, searchFromState)) {
-      store.dispatch(clearCollections())
-      store.dispatch(clearGranules())
-      store.dispatch(clearSelections())
-      store.dispatch(updateSearch(searchFromQuery))
-      store.dispatch(loadData())
-    }
+    store.dispatch(loadCollections(newQueryString))
   }
 }
